CON-451: move method 'equivocatorDetectFromLatestMessage' to EquivocationDetector
